added changes for h2 database

+++ src/main/java/com/telusko/quizapp/model/QuestionWrapper.java
@@ -4,13 +4,13 @@ import lombok.Data;
 
 @Data
 public class QuestionWrapper {
-	private Integer id;
+	private Long id;
     private String QuestionTitle;
     private String option1;
     private String option2;
     private String option3;
     private String option4;
-	public QuestionWrapper(Integer id, String questionTitle, String option1, String option2,
+	public QuestionWrapper(Long id, String questionTitle, String option1, String option2,
 			String option3, String option4) {
 		super();
 		this.id = id;
